LPS-78082 Renaming taglib
