+++ modules/apps/collaboration/blogs/blogs-web/build.gradle
@@ -29,7 +29,7 @@ dependencies {
 	provided project(":apps:collaboration:item-selector:item-selector-criteria-api")
 	provided project(":apps:collaboration:item-selector:item-selector-taglib")
 	provided project(":apps:collaboration:reading-time:reading-time-taglib")
-	provided project(":apps:collaboration:social:social-taglib")
+	provided project(":apps:collaboration:social:social-bookmarks-taglib")
 	provided project(":apps:collaboration:subscription:subscription-api")
 	provided project(":apps:collaboration:upload:upload-api")
 	provided project(":apps:foundation:friendly-url:friendly-url-api")
