+++ modules/apps/collaboration/blogs/blogs-web/src/main/resources/META-INF/resources/blogs/configuration.jsp
@@ -49,7 +49,7 @@ blogsPortletInstanceConfiguration = ConfigurationProviderUtil.getConfiguration(B
 					<liferay-ui:message key="social-bookmarks" />
 				</h4>
 
-				<liferay-social:bookmarks-settings
+				<liferay-social-bookmarks:bookmarks-settings
 					displayStyle="<%= blogsPortletInstanceConfiguration.socialBookmarksDisplayStyle() %>"
 					types="<%= blogsPortletInstanceConfiguration.socialBookmarksTypes() %>"
 				/>
