+++ modules/apps/collaboration/blogs/blogs-web/src/main/resources/META-INF/resources/blogs/social_bookmarks.jsp
@@ -35,7 +35,7 @@ String socialBookmarksDisplayStyle = blogsPortletInstanceConfiguration.socialBoo
 	</c:choose>
 </portlet:renderURL>
 
-<liferay-social:bookmarks
+<liferay-social-bookmarks:bookmarks
 	contentId="<%= String.valueOf(entry.getEntryId()) %>"
 	displayStyle="<%= blogsPortletInstanceConfiguration.socialBookmarksDisplayStyle() %>"
 	target="_blank"
