+++ modules/apps/collaboration/social/social-bookmarks-taglib/build.gradle
