+++ modules/apps/collaboration/social/social-bookmarks-taglib/src/main/resources/META-INF/resources/bookmark/init.jsp
