+++ modules/apps/collaboration/social/social-bookmarks-taglib/src/main/resources/META-INF/resources/bookmarks/init.jsp
