+++ modules/apps/collaboration/social/social-bookmarks-taglib/src/main/resources/META-INF/taglib-mappings.properties
@@ -1,0 +1,1 @@
+liferay_social_bookmarks=/META-INF/liferay-social-bookmarks.tld
