+++ modules/apps/collaboration/social/social-bookmarks-taglib/src/main/resources/com/liferay/social/bookmarks/taglib/servlet/taglib/packageinfo
