--- modules/apps/collaboration/social/social-taglib/bnd.bnd
@@ -1,12 +1,0 @@
-Bundle-Name: Liferay Social Taglib
-Bundle-SymbolicName: com.liferay.social.taglib
-Bundle-Version: 1.0.0
-Export-Package: com.liferay.social.taglib.servlet.taglib
-Liferay-Releng-Module-Group-Description:
-Liferay-Releng-Module-Group-Title: Social
-Provide-Capability:\
-	osgi.extender;\
-		osgi.extender="jsp.taglib";\
-		uri="http://liferay.com/tld/social";\
-		version:Version="${Bundle-Version}"
-Web-ContextPath: /social-taglib
