--- modules/apps/collaboration/social/social-taglib/src/main/resources/META-INF/taglib-mappings.properties
@@ -1,1 +1,0 @@
-liferay_social=/META-INF/liferay-social.tld
