+++ modules/apps/knowledge-base/knowledge-base-web/src/main/resources/META-INF/resources/display/configuration.jsp
@@ -111,7 +111,7 @@ kbDisplayPortletInstanceConfiguration = ParameterMapUtil.setParameterMap(KBDispl
 								<liferay-ui:message key="social-bookmarks" />
 							</h4>
 
-							<liferay-social:bookmarks-settings
+							<liferay-social-bookmarks:bookmarks-settings
 								displayStyle="<%= kbDisplayPortletInstanceConfiguration.socialBookmarksDisplayStyle() %>"
 								types="<%= kbDisplayPortletInstanceConfiguration.socialBookmarksTypes() %>"
 							/>
