+++ modules/apps/web-experience/asset/asset-publisher-web/build.gradle
@@ -28,7 +28,7 @@ dependencies {
 	provided project(":apps:collaboration:comment:comment-taglib")
 	provided project(":apps:collaboration:flags:flags-taglib")
 	provided project(":apps:collaboration:item-selector:item-selector-api")
-	provided project(":apps:collaboration:social:social-taglib")
+	provided project(":apps:collaboration:social:social-bookmarks-taglib")
 	provided project(":apps:collaboration:subscription:subscription-api")
 	provided project(":apps:static:portal-configuration:portal-configuration-metatype-api")
 	provided project(":apps:web-experience:asset:asset-api")
