+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/resources/META-INF/resources/configuration/set_enable.jsp
@@ -133,7 +133,7 @@
 <aui:row cssClass="panel-group">
 	<aui:col>
 		<liferay-ui:panel collapsible="<%= false %>" markupView="lexicon" title="social-bookmarks">
-			<liferay-social:bookmarks-settings
+			<liferay-social-bookmarks:bookmarks-settings
 				displayStyle="<%= assetPublisherDisplayContext.getSocialBookmarksDisplayStyle() %>"
 				types="<%= assetPublisherDisplayContext.getSocialBookmarksTypes() %>"
 			/>
