+++ modules/apps/web-experience/asset/asset-publisher-web/src/main/resources/META-INF/resources/display/full_content.jsp
@@ -138,7 +138,7 @@ request.setAttribute("view.jsp-showIconLabel", true);
 		/>
 
 		<div class="pull-right">
-			<liferay-social:bookmarks
+			<liferay-social-bookmarks:bookmarks
 				displayStyle="<%= assetPublisherDisplayContext.getSocialBookmarksDisplayStyle() %>"
 				target="_blank"
 				title="<%= title %>"
